Move grouping related code to a separate directory under /message
